this commit separate the diff class into separate files for ease of readability
