+++ frontend/src/components/CalendarTimeModal.tsx
@@ -22,7 +22,7 @@ import InfoOutlined from "@mui/icons-material/InfoOutline";
 import AddCircleOutlineIcon from "@mui/icons-material/AddCircleOutline";
 import { EventTimeEnum } from "../utils/constants";
 import { useEventRec } from "../contexts/EventRecContext";
-import TimeBlock from "../../../backend/utils/TimeBlockClass";
+import TimeBlock from "../../../backend/classes/TimeBlock";
 
 type GPEventTimeModal = {
   editMode: boolean;
